Bugfix: wrong sign of A_cont in decomposition for Delta_A_u
Luckily, this had no relevant effect:
Previous stability results are still valid, because the because the used norm-bound is sign-invariant: norm(M1 (expm(A*t) - I) M2) is bounded for |t|<const., so the sign of A is irrelevant.
The only change is in the set of considered Delta_A for the LMI, which was an approximation anyway, so it does *not* invalidate stability results even if we put random garbage there (see report, section 'Verified Numerical Implementation': P is considered only as an unproven guess, which is then checked. The bug only affected P and not the verification.).
However, this slightly changes the heuristics for the LMI, and therefore the determined value of rho (usually in the order of 1e-4). Both the new and the old value are rho are mathematically valid.