Drop Jupyter-Notebook-Server, bump dependencies
Update dependencies to fix broken build. Unfortunately, `pip` is buggy so that even with pinned versions (package==1.2.3 in requirements.txt) the build failed.

Jupyter was only supported by a small subset of the experiments; removing it as an emergency measure to avoid dependency hell and long build times. Can probably be re-added in the future if needed.